+++ crates/core/runner/src/inputs_collector.rs
@@ -112,7 +112,7 @@ pub async fn collect_and_hash_inputs(
     if !task.input_globs.is_empty() {
         // Collect inputs by walking and globbing the file system
         if use_globs {
-            files_to_hash.extend(glob::walk(workspace_root, &task.input_globs)?);
+            files_to_hash.extend(glob::walk_files(workspace_root, &task.input_globs)?);
 
             // Collect inputs by querying VCS
         } else {
@@ -133,7 +133,7 @@ pub async fn collect_and_hash_inputs(
                 .collect::<Vec<_>>();
 
             if !workspace_globs.is_empty() {
-                files_to_hash.extend(glob::walk(workspace_root, workspace_globs)?);
+                files_to_hash.extend(glob::walk_files(workspace_root, workspace_globs)?);
             }
         }
     }
@@ -142,7 +142,13 @@ pub async fn collect_and_hash_inputs(
     // Also run this LAST as it should take highest precedence!
     if !is_ci() {
         for local_file in vcs.get_touched_files().await?.all {
-            files_to_hash.insert(workspace_root.join(local_file));
+            let local_file = workspace_root.join(local_file);
+
+            // Deleted files are listed in `git status` but are
+            // not valid inputs, so avoid hashing them!
+            if local_file.exists() {
+                files_to_hash.insert(local_file);
+            }
         }
     }