+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## Unreleased
+
+#### 🐞 Fixes
+
+- Fixed a few issues during input hashing:
+  - Would attempt to include deleted files from `git status`, which would log a warning.
+  - Would attempt to hash directories for root-level projects, which would log a warning.
+
 ## 1.3.0
 
 #### 🚀 Updates